refactor: extract package service tests
